+++ lib/swiss_knife.dart
@@ -1,4 +1,5 @@
 library swiss_knife;
 
+export 'src/events.dart';
 export 'src/utils.dart';
-export 'src/locales.dart';
+