+++ lib/swiss_knife_browser.dart
@@ -1,7 +1,7 @@
 library swiss_knife;
 
+export 'src/events.dart';
+
 export 'src/utils.dart';
 export 'src/utils_browser.dart';
-export 'src/locales.dart';
-export 'src/locales_browser.dart';
-export 'src/rest_client.dart';
+