+++ lib/swiss_knife_utils.dart
@@ -1,3 +1,0 @@
-library swiss_knife;
-
-export 'src/utils.dart';