+++ lib/swiss_knife_utils_browser.dart
@@ -1,4 +1,0 @@
-library swiss_knife;
-
-export 'src/utils.dart';
-export 'src/utils_browser.dart';